requirements.txt
streamlit==1.28.0
PyPDF2==3.0.1
python-pptx==0.6.21
pytesseract==0.3.10
Pillow==10.0.1
openai==1.3.0
chromadb==0.4.15
python-dotenv==1.0.0
tiktoken==0.5.2
